# Prevent extra `spawn` to make `klass.all` faster