@matthewd 2017-05-08
These generally seem to be mutating scopes that don't belong to the caller. While it's reasonable to use the secret `!` methods internally when we're chaining several operations, I think a few maybe-avoidable allocations are probably better than clouding the "Relation is immutable" rule any more.