@kamipo 2017-08-11
These extra `spawn` are called via `klass.all` and `klass.all` is called everywhere in the internal.
Avoiding the extra `spawn` makes` klass.all` 30% faster for STI classes.
I think it is worth to make `klass.all` faster.

```
Warming up --------------------------------------
       fast relation     4.410k i/100ms
       slow relation     3.334k i/100ms
Calculating -------------------------------------
       fast relation     47.373k (± 5.2%) i/s -    238.140k in   5.041836s
      slow relation     35.757k (±15.9%) i/s -    176.702k in   5.104625s

Comparison:
       fast relation:    47373.2 i/s
       slow relation:    35756.7 i/s - 1.32x  slower
```

```ruby
# frozen_string_literal: true

begin
  require "bundler/inline"
rescue LoadError => e
  $stderr.puts "Bundler version 1.10 or later is required. Please update your Bundler"
  raise e
end

gemfile(true) do
  source "https://rubygems.org"
  gem "rails", github: "rails/rails"
  gem "arel", github: "rails/arel"
  gem "sqlite3"
  gem "benchmark-ips"
end

require "active_record"
require "benchmark/ips"

ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")

ActiveRecord::Schema.define do
  create_table :foos do |t|
    t.string :type
  end
end

class Foo < ActiveRecord::Base
  def self.fast_relation
    relation = ActiveRecord::Relation.create(self, arel_table, predicate_builder)

    if finder_needs_type_condition? && !ignore_default_scope?
      relation.where!(type_condition)
      relation.create_with!(inheritance_column.to_s => sti_name)
    else
      relation
    end
  end
end

class FastFoo < Foo
  class << self
    alias :relation :fast_relation
  end
end

class SlowFoo < Foo
end

GC.start
GC.disable
Benchmark.ips do |x|
  x.report("fast relation") { FastFoo.all }
  x.report("slow relation") { SlowFoo.all }
  x.compare!
end
GC.enable
```