@kamipo 2017-08-21
I included the benchmark result in the commit message.