@kamipo 2017-11-19
I'm going to merge this improvement since 30% faster `klass.all` is used in the internal everywhere so it is worth to all AR users (and I'm always using faster form in the internal).